requirements.txt
requests
pandas
numpy
python-dotenv
openai